Update to OpenGL 4.0, converted vertex coordinates to floats

+++ OpenRR/src/openrr/map/loader/SubTextureCoordinate.java
@@ -4,12 +4,12 @@ import openrr.map.Orientation;
 
 public class SubTextureCoordinate {
 
-	public final double u1;
-	public final double v1;
-	public final double u2;
-	public final double v2;
+	public final float u1;
+	public final float v1;
+	public final float u2;
+	public final float v2;
 
-	public SubTextureCoordinate(double u1, double v1, double u2, double v2) {
+	public SubTextureCoordinate(float u1, float v1, float u2, float v2) {
 		this.u1 = u1;
 		this.v1 = v1;
 		this.u2 = u2;
